ifdef ABD_optim added to g_function and g_function_freq_in.hpp

+++ src/decoders/naive_fixed/f_type.hpp
@@ -6,25 +6,23 @@
 //
 #include "include/ap_fixed.h"
 //
-#define NBITS   26
-#define NINTG   2
+#define NBITS 26
+#define NINTG 2
 //
 struct symbols_f {
     ap_fixed<NBITS, NINTG> value[_GF_];
-    bool is_freq;
+    bool                   is_freq;
 };
 //
 //
 //
 //
 //
-inline float is_issue(const ap_fixed<NBITS, NINTG>* value)
-{
+inline float is_issue(const ap_fixed<NBITS, NINTG> * value) {
     float maxv = 0.f;
-    for (size_t i = 0; i < _GF_; i++)
-    {
+    for (size_t i = 0; i < _GF_; i++) {
         const float v = value[i].to_float();
-        maxv = std::fmaxf( maxv, std::fabsf( v ) );
+        maxv          = std::fmaxf(maxv, std::fabs(v));
     }
     return maxv;
 }
